Added first-pass implementation of Swatch object.
First-pass implementation of the Swatch color data object. Now we return a
data object that has .r, .g, .b, .a, .rgb, .rgba, and .hex values. We use
ES5's .defineProperty (if available) to make them not-writable to prevent
confusion.

Updated tests to use the .hex value to set the background color of cells in
tests.